+++ src/TraceParser.h
@@ -48,7 +48,7 @@
 
 class TraceParser {
  public:
-  TraceParser(int nbrOfBanks);
+  TraceParser(uint64_t nbrOfBanks);
   // list of parsed commands
   std::vector<Data::MemCommand> cmd_list;
